Wait for server startup to avoid 599
This patch avoids a race condition while testing whereby our request
would fire previously to the server actually starting.

With this approach we should have guarantees that our server is up and
running before the Proc::Async passes back to the callsite.

Also, kill the server after we run the tests.